Split site.py into modules.
