+++ site/__main__.py
@@ -1,0 +1,3 @@
+from __init__ import main
+
+main()
